Changed double to float and fix exchangeCoef init

+++ Assets/Code/ResourceMarket/Interfaces/IMarketDataProvider.cs
@@ -5,7 +5,7 @@ namespace ResourceMarket
     public interface IMarketDataProvider
     {
         public event Action<int> OnMarketAmountChange;
-        public double MarketCoef { get; }
+        public float MarketCoef { get; }
         public int MarketAmount { get; }
 
         public void AddMarket();
